Generalized association is now working. Output is in a horrific format. Implementation of T-testing. Improvements are to look for classes dynamically (a la VariantEval/VariantAnnotator), beautify output, and do optimizations where they exist.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@5341 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/oneoffprojects/walkers/association/interfaces/TStatistic.java
@@ -1,5 +1,7 @@
 package org.broadinstitute.sting.oneoffprojects.walkers.association.interfaces;
 
+import org.broadinstitute.sting.utils.collections.Pair;
+
 /**
  * Created by IntelliJ IDEA.
  * User: chartl
@@ -9,5 +11,5 @@ package org.broadinstitute.sting.oneoffprojects.walkers.association.interfaces;
  */
 public interface TStatistic {
     public abstract double getTStatistic();
-    public abstract int getDegreesOfFreedom();
+    public abstract double getDegreesOfFreedom();
 }
